[Webapp] Resolve server-side rendering bug through implementation of new Next..js app directory routing
