+++ src/webapp/package.json
@@ -11,20 +11,20 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
     "antd": "^5.4.5",
-    "eslint-config-next": "^13.3.0",
-    "next": "^13.4.10",
+    "next": "^13.4.19",
     "next-auth": "^4.22.3",
     "node-fetch": "^3.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.1.1"
+    "react-redux": "^8.1.2"
   },
   "devDependencies": {
-    "@types/node": "^18.15.11",
-    "@types/node-fetch": "^2.6.3",
-    "@types/react": "^18.0.34",
-    "@types/react-dom": "^18.0.11",
+    "@types/node": "^20.6.2",
+    "@types/node-fetch": "^2.6.5",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
     "autoprefixer": "^10.4.14",
+    "eslint-config-next": "^13.4.19",
     "postcss": "^8.4.21",
     "prettier": "^2.8.8",
     "sass": "^1.61.0",
