+++ src/webapp/src/app/components/sidebar/utils.ts
