+++ src/webapp/src/app/page.tsx
@@ -1,0 +1,5 @@
+import HomePage from "./HomePage";
+
+export default function Page() {
+  return <HomePage />;
+};
