+++ src/webapp/src/app/profiles/[id]/page.tsx
@@ -1,0 +1,5 @@
+import ProfilePage from "./ProfilePage";
+
+export default function Page() {
+  return <ProfilePage />;
+};
