+++ src/webapp/src/app/providers/index.tsx
@@ -1,0 +1,20 @@
+'use client'
+
+import { Provider } from "react-redux";
+import { SessionProvider } from "next-auth/react";
+
+import { store } from "@/store";
+
+import ThemeProvider from "./themeProvider";
+
+const Providers = ({ children }: React.PropsWithChildren) => {
+  return (
+    <SessionProvider>
+      <Provider store={store}>
+        <ThemeProvider>{children}</ThemeProvider>
+      </Provider>
+    </SessionProvider>
+  );
+};
+
+export default Providers;
