+++ src/webapp/src/app/regions/[id]/page.tsx
@@ -1,0 +1,5 @@
+import RegionPage from "./RegionPage";
+
+export default function Page() {
+  return <RegionPage />;
+};
