+++ src/webapp/src/app/servers/[id]/page.tsx
@@ -1,0 +1,5 @@
+import ServerPage from "./ServerPage";
+
+export default function Page() {
+  return <ServerPage />;
+};
