+++ src/webapp/src/app/servers/page.tsx
@@ -1,0 +1,5 @@
+import ServersPage from "./ServersPage";
+
+export default function Page() {
+  return <ServersPage />;
+};
