+++ src/webapp/src/app/users/[id]/page.tsx
@@ -1,0 +1,5 @@
+import UserPage from "./UserPage";
+
+export default function Page() {
+  return <UserPage />;
+};
