--- src/webapp/src/pages/_document.tsx
@@ -1,15 +1,0 @@
-import { Html, Head, Main, NextScript } from "next/document";
-
-const Document: React.FC = () => {
-  return (
-    <Html lang="en">
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-};
-
-export default Document;
