--- src/webapp/src/pages/components/footer/footer.tsx
@@ -1,17 +1,0 @@
-import React, { Fragment } from "react";
-
-const Footer: React.FC = () => {
-  return (
-    <Fragment>
-      <p>
-        Pokémon Copyright © 1995-2023 Nintendo/Creatures Inc./GAME FREAK Inc
-      </p>
-      <p>
-        PokéSandbox is not affiliated with Nintendo, Creatures Inc. and GAME
-        FREAK Inc.
-      </p>
-    </Fragment>
-  );
-};
-
-export default Footer;
