--- src/webapp/src/pages/servers/index.tsx
@@ -1,30 +1,0 @@
-import React from "react";
-
-import { useGetServers } from "@/store/selectors/serversSelectors";
-
-import ServerCard from "./components/serverCard";
-
-import type { IServer } from "@/interfaces/models/server";
-
-interface ServersProps {
-  servers: IServer[];
-}
-
-const Servers: React.FC<ServersProps> = () => {
-  const servers: IServer[] = useGetServers();
-
-  return (
-    <div>
-      <h1 className="text-center text-xl">Server Discovery</h1>
-      <div className="flex flex-wrap justify-evenly mt-5">
-        {servers.map((server: IServer) =>
-          server.discovery.enabled ? (
-            <ServerCard key={server.serverId} server={server} />
-          ) : null,
-        )}
-      </div>
-    </div>
-  );
-};
-
-export default Servers;
