--- src/webapp/src/providers/index.tsx
@@ -1,22 +1,0 @@
-import { Provider } from "react-redux";
-import { SessionProvider } from "next-auth/react";
-
-import { store } from "@/store";
-
-import ThemeProvider from "./themeProvider";
-
-interface ProvidersProps extends React.PropsWithChildren {
-  session: any;
-}
-
-const Providers = ({ children, session }: ProvidersProps) => {
-  return (
-    <SessionProvider session={session}>
-      <Provider store={store}>
-        <ThemeProvider>{children}</ThemeProvider>
-      </Provider>
-    </SessionProvider>
-  );
-};
-
-export default Providers;
